Refactor. Moved around some UI. DB views in table are broken.
